% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intlforestry-package.R
\docType{package}
\name{intlforestry-package}
\alias{intlforestry}
\alias{intlforestry-package}
\title{intlforestry: Functions for FS Ch.9: Int'l Forestry}
\description{
Reproducible functions and workflows for producing the statistics and text of FS Ch.9: Int'l Forestry.
}
\author{
\strong{Maintainer}: Daniel Braby \email{daniel.braby@forestresearch.gov.uk}

}
\keyword{internal}
